requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
pypdf
openai